Extract code to make it more easily testable
